Conflicts:
	library/payment/method/express/GoogleCheckout.php

+++ build/copyrightJs.sh
@@ -1,4 +1,6 @@
 #!/bin/bash
-cat $1/build/copyright.js $2 > $2.copy
-rm $2
-mv $2.copy $2
+if [ "$2" != "" ]; then
+	cat $1/build/copyright.jss $2 > $2.copy
+	rm $2
+	mv $2.copy $2
+fi
